First EDA Project
Instacart is a grocery delivery platform where customers can place a grocery order and have it delivered to them, similar to how Uber Eats and Door Dash work. This particular dataset was publicly released by Instacart in 2017 for a Kaggle competition. 

Within this notebook I went thru and took the initial data sets that I was given and did and initial EDA on them. I went thru and did things like check for duplicates and missing values. I also went and made different charts so the data could be easily visualized.