ML in Business - New Well Location
In this project we are going to be utilizing various ML techniques to help find the best location for a new oil well. 


I will:
Collect the oil well parameters in the selected region: oil quality and volume of reserves;
Build a model for predicting the volume of reserves in the new wells;
Pick the oil wells with the highest estimated values;
Pick the region with the highest total profit for the selected oil wells.
Build a model that will help to pick the region with the highest profit margin. 
Analyze potential profit and risks using the Bootstrapping technique.