Churn vs No Churn ML Model
The telecom operator Interconnect would like to be able to forecast their churn of clients. If it's discovered that a user is planning to leave, they will be offered promotional codes and special plan options. Interconnect's marketing team has collected some of their clientele's personal data, including information about their plans and contracts.

This is the final project for the TripleTen Bootcamp I did. In this project I built a machine learning model that will predict if a customer will churn or not based off various different factors.